tutorials: my first widget: update "display output" part to split the processing into three files.
