+++ doc/tutorials/my_first_widget/display_output.rst
@@ -112,8 +112,15 @@ Now your processing should looks like:
 .. admonition:: Results
     :class: dropdown
 
-    .. include:: materials/display_output.py
+    .. include:: materials/display_output/clipdata.py
         :literal:
+
+    .. include:: materials/display_output/MyWidget.py
+        :literal:
+
+    .. include:: materials/display_output/OWClipData.py
+        :literal:
+
 
 Further reading
 ---------------
